Update zauberSpiegel.pde
It's not working perfectly. I should come back and fix it. One can use it fine, but the layer with the movies and and the one with the black frame are not working.
You can download them already if you want though:
the frame: https://cloud.medienhaus.udk-berlin.de/f/22ddeabc41/?raw=1
the movie loop with the closes: https://cloud.medienhaus.udk-berlin.de/f/ed0103c1c6/?raw=1